requirements.txt
PyGithub
json-tricks